Moving cdrom files to their own project.
