+++ src/cdrom/cdriso-ecm.cc
@@ -17,7 +17,7 @@
  *   51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.           *
  ***************************************************************************/
 
-#include "core/cdriso.h"
+#include "cdrom/cdriso.h"
 #include "core/cdrom.h"
 
 /* Adapted from ecm.c:unecmify() (C) Neill Corlett */
