+++ src/cdrom/cdriso.h
@@ -23,8 +23,8 @@
 
 #include <filesystem>
 
-#include "core/iec-60908b.h"
-#include "core/ppf.h"
+#include "cdrom/iec-60908b.h"
+#include "cdrom/ppf.h"
 #include "core/psxemulator.h"
 #include "support/uvfile.h"
