+++ src/cdrom/iec-60908b.cc
@@ -17,7 +17,7 @@
  *   51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.           *
  ***************************************************************************/
 
-#include "core/iec-60908b.h"
+#include "cdrom/iec-60908b.h"
 
 #include <stdint.h>
