+++ src/cdrom/ppf.h
@@ -21,7 +21,7 @@
 
 #include <stdint.h>
 
-#include "core/iec-60908b.h"
+#include "cdrom/iec-60908b.h"
 
 namespace PCSX {
