+++ vsprojects/cdrom/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="FFmpeg.Nightly" version="20200831.1.0" targetFramework="native" />
+</packages>
