+++ vsprojects/core/packages.config
@@ -1,6 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="FFmpeg.Nightly" version="20200831.1.0" targetFramework="native" />
   <package id="glfw" version="3.3.5" targetFramework="native" />
   <package id="luajit.native" version="2.1.0-beta3b" targetFramework="native" />
 </packages>
