+++ vsprojects/pcsx-redux.sln
@@ -34,7 +34,7 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "libuv", "libuv\libuv.vcxpro
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{9D5A1DB2-E74D-4CDD-8377-9EA08CF4AADE}"
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "support", "tests\support\testsupport.vcxproj", "{50EB9A3C-CC0B-4D74-BCF3-8E6C8B778E5A}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "testsupport", "tests\support\testsupport.vcxproj", "{50EB9A3C-CC0B-4D74-BCF3-8E6C8B778E5A}"
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "memoryleakdetector", "tests\memoryleakdetector\memoryleakdetector.vcxproj", "{DD5ACB0A-E326-4EA9-B5A8-C23D66C27650}"
 EndProject
@@ -77,6 +77,8 @@ EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "gtest", "gtest\gtest.vcxproj", "{432D6160-7127-4005-BFA6-7C301C0CF3D3}"
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "support", "support\support.vcxproj", "{0E621321-093C-4D60-BD8B-027FDC2B0F63}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "cdrom", "cdrom\cdrom.vcxproj", "{026AECDD-EB41-4AFD-866C-59F9FE886FF6}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -487,8 +489,8 @@ Global
 		{5B01D900-2359-44CA-9914-6B0C6AFB7BE7}.ReleaseWithClangCL|x86.Build.0 = ReleaseWithClangCL|Win32
 		{5B01D900-2359-44CA-9914-6B0C6AFB7BE7}.ReleaseWithTracy|x64.ActiveCfg = Release|x64
 		{5B01D900-2359-44CA-9914-6B0C6AFB7BE7}.ReleaseWithTracy|x64.Build.0 = Release|x64
-		{5B01D900-2359-44CA-9914-6B0C6AFB7BE7}.ReleaseWithTracy|x86.ActiveCfg = Release|Win32
-		{5B01D900-2359-44CA-9914-6B0C6AFB7BE7}.ReleaseWithTracy|x86.Build.0 = Release|Win32
+		{5B01D900-2359-44CA-9914-6B0C6AFB7BE7}.ReleaseWithTracy|x86.ActiveCfg = ReleaseWithTracy|Win32
+		{5B01D900-2359-44CA-9914-6B0C6AFB7BE7}.ReleaseWithTracy|x86.Build.0 = ReleaseWithTracy|Win32
 		{A057157E-7638-474A-9D02-91483F20B301}.Debug|x64.ActiveCfg = Debug|x64
 		{A057157E-7638-474A-9D02-91483F20B301}.Debug|x64.Build.0 = Debug|x64
 		{A057157E-7638-474A-9D02-91483F20B301}.Debug|x86.ActiveCfg = Debug|Win32
@@ -503,8 +505,8 @@ Global
 		{A057157E-7638-474A-9D02-91483F20B301}.ReleaseWithClangCL|x86.Build.0 = ReleaseWithClangCL|Win32
 		{A057157E-7638-474A-9D02-91483F20B301}.ReleaseWithTracy|x64.ActiveCfg = Release|x64
 		{A057157E-7638-474A-9D02-91483F20B301}.ReleaseWithTracy|x64.Build.0 = Release|x64
-		{A057157E-7638-474A-9D02-91483F20B301}.ReleaseWithTracy|x86.ActiveCfg = Release|Win32
-		{A057157E-7638-474A-9D02-91483F20B301}.ReleaseWithTracy|x86.Build.0 = Release|Win32
+		{A057157E-7638-474A-9D02-91483F20B301}.ReleaseWithTracy|x86.ActiveCfg = ReleaseWithTracy|Win32
+		{A057157E-7638-474A-9D02-91483F20B301}.ReleaseWithTracy|x86.Build.0 = ReleaseWithTracy|Win32
 		{25C13988-A8A8-4BFA-962F-0833020E4EE4}.Debug|x64.ActiveCfg = Debug|x64
 		{25C13988-A8A8-4BFA-962F-0833020E4EE4}.Debug|x64.Build.0 = Debug|x64
 		{25C13988-A8A8-4BFA-962F-0833020E4EE4}.Debug|x86.ActiveCfg = Debug|Win32
@@ -553,6 +555,22 @@ Global
 		{0E621321-093C-4D60-BD8B-027FDC2B0F63}.ReleaseWithTracy|x64.Build.0 = ReleaseWithTracy|x64
 		{0E621321-093C-4D60-BD8B-027FDC2B0F63}.ReleaseWithTracy|x86.ActiveCfg = ReleaseWithTracy|Win32
 		{0E621321-093C-4D60-BD8B-027FDC2B0F63}.ReleaseWithTracy|x86.Build.0 = ReleaseWithTracy|Win32
+		{026AECDD-EB41-4AFD-866C-59F9FE886FF6}.Debug|x64.ActiveCfg = Debug|x64
+		{026AECDD-EB41-4AFD-866C-59F9FE886FF6}.Debug|x64.Build.0 = Debug|x64
+		{026AECDD-EB41-4AFD-866C-59F9FE886FF6}.Debug|x86.ActiveCfg = Debug|Win32
+		{026AECDD-EB41-4AFD-866C-59F9FE886FF6}.Debug|x86.Build.0 = Debug|Win32
+		{026AECDD-EB41-4AFD-866C-59F9FE886FF6}.Release|x64.ActiveCfg = Release|x64
+		{026AECDD-EB41-4AFD-866C-59F9FE886FF6}.Release|x64.Build.0 = Release|x64
+		{026AECDD-EB41-4AFD-866C-59F9FE886FF6}.Release|x86.ActiveCfg = Release|Win32
+		{026AECDD-EB41-4AFD-866C-59F9FE886FF6}.Release|x86.Build.0 = Release|Win32
+		{026AECDD-EB41-4AFD-866C-59F9FE886FF6}.ReleaseWithClangCL|x64.ActiveCfg = ReleaseWithClangCL|x64
+		{026AECDD-EB41-4AFD-866C-59F9FE886FF6}.ReleaseWithClangCL|x64.Build.0 = ReleaseWithClangCL|x64
+		{026AECDD-EB41-4AFD-866C-59F9FE886FF6}.ReleaseWithClangCL|x86.ActiveCfg = ReleaseWithClangCL|Win32
+		{026AECDD-EB41-4AFD-866C-59F9FE886FF6}.ReleaseWithClangCL|x86.Build.0 = ReleaseWithClangCL|Win32
+		{026AECDD-EB41-4AFD-866C-59F9FE886FF6}.ReleaseWithTracy|x64.ActiveCfg = Release|x64
+		{026AECDD-EB41-4AFD-866C-59F9FE886FF6}.ReleaseWithTracy|x64.Build.0 = Release|x64
+		{026AECDD-EB41-4AFD-866C-59F9FE886FF6}.ReleaseWithTracy|x86.ActiveCfg = ReleaseWithTracy|Win32
+		{026AECDD-EB41-4AFD-866C-59F9FE886FF6}.ReleaseWithTracy|x86.Build.0 = ReleaseWithTracy|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -588,9 +606,10 @@ Global
 		{25C13988-A8A8-4BFA-962F-0833020E4EE4} = {64A05F50-3203-42CC-B632-09D6EE6EA856}
 		{432D6160-7127-4005-BFA6-7C301C0CF3D3} = {64A05F50-3203-42CC-B632-09D6EE6EA856}
 		{0E621321-093C-4D60-BD8B-027FDC2B0F63} = {008A2872-432F-480B-828D-FF9AAA4846BC}
+		{026AECDD-EB41-4AFD-866C-59F9FE886FF6} = {008A2872-432F-480B-828D-FF9AAA4846BC}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		                SolutionGuid = {284E91C1-764E-441A-854D-CFD3623A6501}
 		SolutionGuid = {AC54A867-F976-4B3D-A6EF-F57EB764DCD4}
+		                SolutionGuid = {284E91C1-764E-441A-854D-CFD3623A6501}
 	EndGlobalSection
 EndGlobal
